move loading container
