+++ src/components/pages/rep.tsx
@@ -2,7 +2,7 @@ import { BaseRecord, IResourceComponentsProps, useList } from "@refinedev/core";
 import { Row, Col, Card, List } from "antd";
 import { useSearchParamsState } from "../../utils";
 import { ChartPieRepCollecte } from "../chart_pie_rep_collecte";
-import { LoadingComponent } from "../loading_container";
+import { LoadingComponent } from "../../g2f-dashboard/components/loading_container";
 import { Attribution } from "../attributions";
 import { RepTopbar } from "../rep_topbar";
 import alasql from "alasql";
