+++ src/components/pages/rep_pa.tsx
@@ -3,7 +3,7 @@ import { useSearchParamsState } from "../../utils"
 import { Row, Col, Card } from "antd"
 import { Attribution } from "../attributions"
 import { ChartPieRepCollecte } from "../chart_pie_rep_collecte"
-import { LoadingComponent } from "../loading_container"
+import { LoadingComponent } from "../../g2f-dashboard/components/loading_container"
 import { RepTopbar } from "../rep_topbar"
 import { useState } from "react"
 import { ChartEvolutionRepCollecte } from "../chart_evolution_rep_collecte"
